Refactor for a better separation of concerns.
The BotCore class now maintains the Channels objects separate from
the Connection objects. This allows for a relatively simple
message-passing scheme in which a message is received by a Channel,
the channel hands it off to the bot which then hands it off to the
various modules. If the modules want to speak in the channel that
spawned the message, they would just use the reference they get.
If the module wanted to send a message to all of the channels it
could "yell" using the BotCore object reference.